remove title snapshot test for action menu
